+++ mail/mail-uniffi/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2025-03-xx
 
+### Added
+
+  - `AttachmentList::remove`: Remove attachments from a draft.
+
+### Fixed
+
+  - Fix save issue when attachment are present in a reply/forward draft.
+
 ## [0.60.0] - 2025-03-20
 
 ### Added
@@ -17,7 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `is_valid_email_address`: Validate whether an email address is correctly formatted.
   - `report_an_issue` method for performing API report
   - `IssueReport` struct for creating a report
-  - `AttachmentList::remove`: Remove attachments from a draft.
 
 ### Fixed